+++ bin/irregular/rebuild_mongo.dart
@@ -151,18 +151,18 @@ Future<void> recreateDaCongestionCompactIsone() async {
 //   await archive.dbConfig.db.close();
 // }
 
-Future<void> recreateDaEnergyOffersIsone() async {
-  var archive = DaEnergyOfferArchive();
-  await archive.setupDb();
-  var files = Directory(archive.dir).listSync().whereType<File>().toList();
-  files.sort((a, b) => a.path.compareTo(b.path));
-  await archive.dbConfig.db.open();
-  for (var file in files) {
-    var data = archive.processFile(file);
-    await archive.insertData(data);
-  }
-  await archive.dbConfig.db.close();
-}
+// Future<void> recreateDaEnergyOffersIsone() async {
+//   var archive = DaEnergyOfferArchive();
+//   await archive.setupDb();
+//   var files = Directory(archive.dir).listSync().whereType<File>().toList();
+//   files.sort((a, b) => a.path.compareTo(b.path));
+//   await archive.dbConfig.db.open();
+//   for (var file in files) {
+//     var data = archive.processFile(file);
+//     await archive.insertData(data);
+//   }
+//   await archive.dbConfig.db.close();
+// }
 
 Future<void> recreateFwdResAuctionResults() async {
   var archive = FwdResAuctionResultsArchive();